# monolith/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for GoWallet",
        "title": "GoWallet Monolith API",
        "contact": {
            "name": "API Support",
            "email": "bashocode@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/users": {
            "get": {
                "description": "Get all users from database with pagination (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Get All Users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page limit (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort column (default: created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort order (default: desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_user_model.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/ledger/mutations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all mutations (credit/debit ledger entries) for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger"
                ],
                "summary": "Get Mutation History",
                "responses": {
                    "200": {
                        "description": "Returns success: true and data: []model.LedgerEntry",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/ledger/reconcile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify if the current wallet balance matches the sum of ledger mutations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger"
                ],
                "summary": "Reconcile Wallet Balance",
                "responses": {
                    "200": {
                        "description": "Returns success: true, is_consistent: bool, wallet_balance: decimal.Decimal, calculated_balance: decimal.Decimal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/transactions/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated list of transactions involving the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Transaction History",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page limit (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort column (default: created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort order (default: desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by status (success/failed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_transaction_model.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/transactions/topup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Top up authenticated user's own wallet balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Top Up Wallet Balance",
                "parameters": [
                    {
                        "description": "topup payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_transaction_model.TopUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success: true, message: Success, and data: model.Transaction",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/transactions/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfer money to another user's wallet using email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transfer Balance",
                "parameters": [
                    {
                        "description": "transfer payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_transaction_model.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success: true, message: Success, and data: model.Transaction",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bashocode_gowallet_monolith_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_bashocode_gowallet_monolith_internal_transaction_model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_transaction_model.PaginationMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_bashocode_gowallet_monolith_internal_transaction_model.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "github_com_bashocode_gowallet_monolith_internal_transaction_model.TopUpRequest": {
            "type": "object",
            "required": [
                "amount",
                "idempotency_key"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 100000
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "unique-uuid-key-abc"
                }
            }
        },
        "github_com_bashocode_gowallet_monolith_internal_transaction_model.TransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "idempotency_key",
                "receiver_email"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 50000
                },
                "description": {
                    "type": "string",
                    "example": "Dinner split"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "unique-uuid-key-123"
                },
                "receiver_email": {
                    "type": "string",
                    "example": "receiver@example.com"
                }
            }
        },
...
        "github_com_bashocode_gowallet_monolith_internal_user_model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/github_com_bashocode_gowallet_monolith_internal_user_model.PaginationMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_bashocode_gowallet_monolith_internal_user_model.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003cyour_token\u003e\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}